fix(desktop): avoid AppImage EGL startup failure
